+++ warp/tests/test_generics.py
@@ -554,15 +554,15 @@ mat22d = wp.types.matrix((2, 2), wp.float64)
 
 @wp.kernel
 def test_annotations_kernel():
-    vi16 = wp.vec3s(wp.int16(1), wp.int16(2), wp.int16(3))
-    vf64 = wp.vec3d(wp.float64(1), wp.float64(2), wp.float64(3))
+    vi16 = wp.vec3s(1, 2, 3)
+    vf64 = wp.vec3d(1, 2, 3)
     wp.expect_eq(vec_int_annotation_func(vi16), wp.int16(6))
     wp.expect_eq(vec_float_annotation_func(vf64), wp.float64(6))
     wp.expect_eq(vec_scalar_annotation_func(vi16), wp.int16(6))
     wp.expect_eq(vec_scalar_annotation_func(vf64), wp.float64(6))
 
-    mi16 = mat22s(wp.int16(1), wp.int16(2), wp.int16(3), wp.int16(4))
-    mf64 = mat22d(wp.float64(1), wp.float64(2), wp.float64(3), wp.float64(4))
+    mi16 = mat22s(1, 2, 3, 4)
+    mf64 = mat22d(1, 2, 3, 4)
     wp.expect_eq(mat_int_annotation_func(mi16), wp.int16(10))
     wp.expect_eq(mat_float_annotation_func(mf64), wp.float64(10))
     wp.expect_eq(mat_scalar_annotation_func(mi16), wp.int16(10))
